Refactored validator.go to make public the individual validation routines.
